requirements.txt
websocket-client
coloroma